Reorganized source folders and changed namespace
