+++ composer.json
@@ -23,7 +23,7 @@
     },
     "autoload": {
         "psr-4": {
-            "Laasti\\": "src/"
+            "Laasti\\Mailer\\": "src/"
         }
     },
     "autoload-dev": {
