+++ src/Exceptions/SendException.php
@@ -20,21 +20,12 @@
  * Created by PhpStorm.
  * User: msowers
  * Date: 3/30/15
- * Time: 2:42 PM
+ * Time: 1:47 PM
  */
 
-namespace Tx\Mailer\Exceptions;
+namespace Laasti\Mailer\Exceptions;
 
-
-class CodeException extends SMTPException
+class SendException extends SMTPException
 {
-    public function __construct($expected, $received, $serverMessage = null)
-    {
-        $message = "Unexpected return code - Expected: {$expected}, Got: {$received}";
-        if (isset($serverMessage)) {
-            $message .= " | " . $serverMessage;
-        }
-        parent::__construct($message);
-    }
 
 }
