+++ src/Mailer.php
@@ -1,4 +1,4 @@
-<?php namespace Tx;
+<?php namespace Laasti\Mailer;
 /***************************************************\
  *
  *  Mailer (https://github.com/txthinking/Mailer)
@@ -17,14 +17,14 @@
  *
  \***************************************************/
 
-use Tx\Mailer\Message;
-use Tx\Mailer\Servers\ServerInterface;
+use Laasti\Mailer\Message;
+use Laasti\Mailer\Servers\ServerInterface;
 
 /**
  * Class Mailer
  *
  * This class provides the Mailer public methods for backwards compatibility, but it is recommended
- * that you use the Tx\Mailer\SMTP and Tx\Mailer\Message classes going forward
+ * that you use the Laasti\Mailer\SMTP and Laasti\Mailer\Message classes going forward
  *
  * @package Tx
  */
